# backend/docs/swagger.yaml
basePath: /
definitions:
...
  internal_api_rest_v1.PostListItem:
    properties:
      created_at:
        type: string
      id:
        type: integer
      qid:
        type: string
      title:
        type: string
    type: object
...
  internal_api_rest_v1.PostsListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_api_rest_v1.PostListItem'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  markpost_internal_apierr.ErrorResponse:
    properties:
      code:
        type: string
      errors:
        items:
          $ref: '#/definitions/markpost_internal_apierr.FieldError'
        type: array
      message:
        type: string
    type: object
  markpost_internal_apierr.FieldError:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
...
host: localhost:7330
info:
  contact: {}
  description: Markpost backend API for post management and delivery.
  title: Markpost API
  version: "1.0"
paths:
...
  /api/v1/admin/posts/{id}:
    delete:
      parameters:
      - description: Post QID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete any post (admin)
      tags:
      - admin
...
  /api/v1/posts:
    get:
      parameters:
      - description: Title/body search (B3.3)
        in: query
        name: search
        type: string
      - default: 1
        description: Page number (min 1)
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (min 1)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_rest_v1.PostsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List the current user's posts
      tags:
      - posts
  /api/v1/posts/{id}:
    delete:
      parameters:
      - description: Post QID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a post owned by the current user
      tags:
      - posts
securityDefinitions:
  BearerAuth:
    description: '"Bearer <access_token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"